Added support for legacy Nvidia drivers.
Not sure if this works on 96xx, but it works for 173xx.
